requirements.txt
PyPDF2==3.0.1
duckduckgo-search==5.3.0b3
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
streamlit==1.32.2
pandas==2.2.1
openpyxl==3.1.2
PyMuPDF==1.23.26
google-generativeai==0.4.1
